SDR-331 refactor: 테스트코드 패키지 이름 변경 aws -> image
